refactor: hw_hal 全量 platform driver 化 + board 内存微优化
6 个 hw_hal 模块 (gpio/i2c/spi/i2s/uart/rmt_led) 添加 DRIVER_REGISTER,
上层消费者 (st7789/mp3/serial) 改为 device_write/device_ioctl 调用。

board: device_node_t 字段 int→uint8_t 重排 (-8 bytes/节点),
board_driver ok/fail int→uint8_t, task_config core_id int→int8_t.

+++ components/board/include/task_config.h
@@ -12,7 +12,7 @@ typedef struct
     const char* name;
     uint32_t stack_size;
     uint32_t priority;
-    int core_id;
+    int8_t core_id;
 } board_task_config_t;
 
 extern const board_task_config_t board_task_ui;
